# services/API_DOCUMENTATION.yaml
openapi: 3.0.3
info:
  title: Apollo Platform API
  description: |
    ## Apollo Criminal Investigation Platform - Comprehensive API Documentation

    The Apollo Platform provides a complete suite of APIs for criminal investigation,
    intelligence gathering, and threat analysis. This documentation covers all available
    endpoints across Node.js microservices and Python intelligence services.

    ### Service Architecture
    - **Node.js Microservices** (Port 3000-3010): Authentication, Operations, Intelligence, Search, Analytics, Notifications
    - **Python Intelligence Server** (Port 8000): OSINT, Blockchain, Facial Recognition, Voice Analysis, GEOINT

    ### Authentication Methods
    - **JWT Bearer Token**: Primary authentication for all protected endpoints
    - **API Key**: Alternative authentication via `X-API-Key` header
    - **MFA**: Multi-factor authentication for sensitive operations

    ### Rate Limiting
    - Free tier: 100 requests/hour
    - Pro tier: 1,000 requests/hour
    - Enterprise: 10,000 requests/hour

    ### Error Handling
    All errors follow a consistent format with error codes and messages.

  version: 2.0.0
  contact:
    name: Apollo Platform Team
    email: dev@apollo-platform.com
    url: https://apollo-platform.com/support
  license:
    name: Proprietary
    url: https://apollo-platform.com/license
  x-logo:
    url: https://apollo-platform.com/logo.png
    altText: Apollo Platform
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication
  schemas:
...
    Alert:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [security, intelligence, operation, system, compliance, facial_match, transaction, surveillance]
        severity:
          type: string
          enum: [critical, high, medium, low, info]
        title:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [new, acknowledged, in_progress, resolved, dismissed]
        source:
          type: string
        relatedEntityType:
          type: string
        relatedEntityId:
          type: string
        actionRequired:
          type: boolean
        assignedTo:
          type: string
          format: uuid
        acknowledgedAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time

    # ============ OSINT Schemas ============
...
paths:
...
  /api/alerts:
    get:
      summary: List alerts
      description: Get all alerts with filtering
      tags: [Alerts]
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [new, acknowledged, in_progress, resolved, dismissed]
        - name: severity
          in: query
          schema:
            type: string
            enum: [critical, high, medium, low, info]
        - name: type
          in: query
          schema:
            type: string
            enum: [security, intelligence, operation, system, compliance, facial_match, transaction, surveillance]
        - name: assignedTo
          in: query
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [created_at, severity, status, type]
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: List of alerts
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      alerts:
                        type: array
                        items:
                          $ref: '#/components/schemas/Alert'
                      pagination:
                        type: object
    post:
      summary: Create alert
      description: Create a new alert
      tags: [Alerts]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, severity, title, message]
              properties:
                type:
                  type: string
                  enum: [security, intelligence, operation, system, compliance, facial_match, transaction, surveillance]
                severity:
                  type: string
                  enum: [critical, high, medium, low, info]
                title:
                  type: string
                message:
                  type: string
                source:
                  type: string
                relatedEntityType:
                  type: string
                relatedEntityId:
                  type: string
                actionRequired:
                  type: boolean
                  default: false
                assignedTo:
                  type: string
                  format: uuid
                metadata:
                  type: object
      responses:
        '201':
          description: Alert created
  /api/alerts/{id}:
    get:
      summary: Get alert
      description: Get alert by ID
      tags: [Alerts]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Alert details
        '404':
          description: Alert not found
  /api/alerts/{id}/acknowledge:
    patch:
      summary: Acknowledge alert
      description: Acknowledge an alert
      tags: [Alerts]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  format: uuid
                notes:
                  type: string
      responses:
        '200':
          description: Alert acknowledged
  /api/alerts/{id}/progress:
    patch:
      summary: Set alert in progress
      description: Mark alert as being worked on
      tags: [Alerts]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  format: uuid
                notes:
                  type: string
      responses:
        '200':
          description: Alert in progress
  /api/alerts/{id}/resolve:
    patch:
      summary: Resolve alert
      description: Mark alert as resolved
      tags: [Alerts]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  format: uuid
                resolution:
                  type: string
                notes:
                  type: string
      responses:
        '200':
          description: Alert resolved
  /api/alerts/{id}/dismiss:
    patch:
      summary: Dismiss alert
      description: Dismiss an alert
      tags: [Alerts]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  format: uuid
                reason:
                  type: string
      responses:
        '200':
          description: Alert dismissed
  /api/alerts/{id}/assign:
    patch:
      summary: Assign alert
      description: Assign alert to a user
      tags: [Alerts]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assignedTo]
              properties:
                assignedTo:
                  type: string
                  format: uuid
                assignedBy:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Alert assigned
  /api/alerts/stats/summary:
    get:
      summary: Alert statistics
      description: Get alert statistics summary
      tags: [Alerts]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Alert statistics
  /api/alerts/bulk/acknowledge:
    post:
      summary: Bulk acknowledge alerts
      description: Acknowledge multiple alerts at once
      tags: [Alerts]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [alertIds, userId]
              properties:
                alertIds:
                  type: array
                  items:
                    type: string
                    format: uuid
                userId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Alerts acknowledged
  /api/alerts/cleanup:
    delete:
      summary: Cleanup old alerts
      description: Delete old resolved/dismissed alerts
      tags: [Alerts]
      security:
        - bearerAuth: []
      parameters:
        - name: olderThanDays
          in: query
          schema:
            type: integer
            default: 30
      responses:
        '200':
          description: Alerts cleaned up

  # ============================================================
  # PYTHON INTELLIGENCE SERVER ENDPOINTS
  # ============================================================

  # ============ OSINT ENDPOINTS ============
...
security:
  - bearerAuth: []
  - apiKeyAuth: []